docs/08-code-and-snippets/20-templates/project-skeleton-python/requirements.txt
openai==1.13.3
python-dotenv==1.0.1
tiktoken==0.6.0
httpx==0.27.0
